+++ src/main/java/morty/command/DeleteCommand.java
@@ -1,44 +1,42 @@
-package duke.command;
+package morty.command;
 
-import duke.TaskList;
-
-import duke.Storage;
-import duke.Ui;
-import duke.task.Task;
+import morty.Storage;
+import morty.TaskList;
+import morty.Ui;
+import morty.task.Task;
 
 /**
  * Represents a command to add a deadline task.
  */
-public class DoneCommand extends Command {
+public class DeleteCommand extends Command {
 
   /**
-   * Constructs a DoneCommand object with the given tokens.
+   * Constructs a DeleteCommand object with the given tokens.
    *
    * @param tokens The tokens of the user input.
    */
-  public DoneCommand(String[] tokens) {
+  public DeleteCommand(String[] tokens) {
     super(tokens);
   }
 
   /**
-   * Executes the DoneCommand.
+   * Executes the DeleteCommand.
    *
    * @param tasks   The list of tasks.
    * @param ui      The user interface.
    * @param storage The storage.
    */
-  @Override
   public void execute(TaskList tasks, Ui ui, Storage storage) {
     try {
       int index = Integer.parseInt(tokens[1]) - 1;
       Task task = tasks.get(index);
-      tasks.markDone(index);
-      ui.showTaskMarkedDone(task);
+      tasks.remove(index);
+      ui.showTaskRemoved(task, tasks.getSize());
       storage.save(tasks);
     } catch (NumberFormatException e) {
-      ui.showInvalidDone();
+      ui.showInvalidDelete();
     } catch (IndexOutOfBoundsException e) {
-      ui.showInvalidDone();
+      ui.showInvalidDelete();
     }
   }
 }
