+++ src/main/java/morty/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package morty.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
